+++ kernel/core/loader.c
@@ -98,7 +98,7 @@ task_entry load_elf (char *elf_start, int idx)
 
     /* to do page directory construction here */
     /* ismi*/
-    unsigned long *ppd = 0x0;
+    unsigned int *ppd = 0x0;
 
 /* user task is not inserted to rq. it should be called explicitly*/
     sprintf(buff,"user%d",idx);
